refactor(server): update entity module service and resolver
